+++ sdk/compute/azure-mgmt-avs/generated_samples/private_clouds_delete.py
@@ -29,13 +29,12 @@ def main():
         subscription_id="00000000-0000-0000-0000-000000000000",
     )
 
-    response = client.private_clouds.begin_delete(
+    client.private_clouds.begin_delete(
         resource_group_name="group1",
         private_cloud_name="cloud1",
     ).result()
-    print(response)
 
 
-# x-ms-original-file: specification/vmware/resource-manager/Microsoft.AVS/stable/2022-05-01/examples/PrivateClouds_Delete.json
+# x-ms-original-file: specification/vmware/resource-manager/Microsoft.AVS/stable/2023-03-01/examples/PrivateClouds_Delete.json
 if __name__ == "__main__":
     main()