+++ sdk/ml/azure-ai-ml/azure/ai/ml/_ml_client.py
@@ -2,7 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
-# pylint: disable=client-accepts-api-version-keyword,too-many-statements,too-many-instance-attributes,too-many-lines,using-constant-test
+# pylint: disable=client-accepts-api-version-keyword,(line-too-long,too-many-locals,no-member,too-many-statements,too-many-instance-attributes,too-many-lines,using-constant-test
 
 import json
 import logging
@@ -164,7 +164,8 @@ class MLClient:
             )
 
         self._credential = credential
-
+        self._ws_rg = None
+        self._ws_sub = None
         show_progress = kwargs.pop("show_progress", True)
         enable_telemetry = kwargs.pop("enable_telemetry", True)
         self._operation_config = OperationConfig(show_progress=show_progress, enable_telemetry=enable_telemetry)
@@ -199,12 +200,53 @@ class MLClient:
         # registry_name is present when the operations need referring assets from registry.
         # the subscription, resource group, if provided, will be ignored and replaced by
         # whatever is received from the registry discovery service.
-        if registry_name:
-            self._service_client_10_2021_dataplanepreview, resource_group_name, subscription_id = get_registry_client(
-                self._credential, registry_name, **kwargs
+        workspace_location = None
+        workspace_id = None
+        registry_reference = kwargs.pop("registry_reference", None)
+        if registry_name or registry_reference:
+            # get the workspace location here if workspace_reference is provided
+            self._ws_operation_scope = OperationScope(
+                subscription_id,
+                resource_group_name,
+                workspace_name,
             )
+            workspace_reference = kwargs.pop("workspace_reference", None)
+            if workspace_reference or registry_reference:
+                ws_ops = WorkspaceOperations(
+                    OperationScope(subscription_id, resource_group_name, workspace_reference),
+                    ServiceClient042023Preview(
+                        credential=self._credential,
+                        subscription_id=subscription_id,
+                        **kwargs,
+                    ),
+                    self._credential,
+                )
+                self._ws_rg = resource_group_name
+                self._ws_sub = subscription_id
+                workspace_details = ws_ops.get(workspace_reference if workspace_reference else workspace_name)
+                workspace_location, workspace_id = (
+                    workspace_details.location,
+                    workspace_details._workspace_id,
+                )
 
-        self._operation_scope = OperationScope(subscription_id, resource_group_name, workspace_name, registry_name)
+            (
+                self._service_client_10_2021_dataplanepreview,
+                resource_group_name,
+                subscription_id,
+            ) = get_registry_client(
+                self._credential, registry_name if registry_name else registry_reference, workspace_location, **kwargs
+            )
+            if not workspace_name:
+                workspace_name = workspace_reference
+
+        self._operation_scope = OperationScope(
+            subscription_id,
+            resource_group_name,
+            workspace_name,
+            registry_name,
+            workspace_id,
+            workspace_location,
+        )
 
         # Cannot send multiple base_url as azure-cli sets the base_url automatically.
         kwargs.pop("base_url", None)
@@ -222,7 +264,9 @@ class MLClient:
         user_agent = kwargs.get("user_agent", None)
 
         app_insights_handler = get_appinsights_log_handler(
-            user_agent, **{"properties": properties}, enable_telemetry=self._operation_config.enable_telemetry
+            user_agent,
+            **{"properties": properties},
+            enable_telemetry=self._operation_config.enable_telemetry,
         )
         app_insights_handler_kwargs = {"app_insights_handler": app_insights_handler}
 
@@ -260,7 +304,9 @@ class MLClient:
 
         self._service_client_04_2023 = ServiceClient042023(
             credential=self._credential,
-            subscription_id=self._operation_scope._subscription_id,
+            subscription_id=self._ws_operation_scope._subscription_id
+            if registry_reference
+            else self._operation_scope._subscription_id,
             base_url=base_url,
             **kwargs,
         )
@@ -278,7 +324,9 @@ class MLClient:
 
         self._service_client_10_2022_preview = ServiceClient102022Preview(
             credential=self._credential,
-            subscription_id=self._operation_scope._subscription_id,
+            subscription_id=self._ws_operation_scope._subscription_id
+            if registry_reference
+            else self._operation_scope._subscription_id,
             base_url=base_url,
             **kwargs,
         )
@@ -299,20 +347,24 @@ class MLClient:
 
         self._service_client_04_2023_preview = ServiceClient042023Preview(
             credential=self._credential,
-            subscription_id=self._operation_scope._subscription_id,
+            subscription_id=self._ws_operation_scope._subscription_id
+            if registry_reference
+            else self._operation_scope._subscription_id,
             base_url=base_url,
             **kwargs,
         )
 
         self._service_client_06_2023_preview = ServiceClient062023Preview(
             credential=self._credential,
-            subscription_id=self._operation_scope._subscription_id,
+            subscription_id=self._ws_operation_scope._subscription_id
+            if registry_reference
+            else self._operation_scope._subscription_id,
             base_url=base_url,
             **kwargs,
         )
 
         self._workspaces = WorkspaceOperations(
-            self._operation_scope,
+            self._ws_operation_scope if registry_reference else self._operation_scope,
             self._service_client_06_2023_preview,
             self._operation_container,
             self._credential,
@@ -368,15 +420,21 @@ class MLClient:
         self._models = ModelOperations(
             self._operation_scope,
             self._operation_config,
-            self._service_client_10_2021_dataplanepreview if registry_name else self._service_client_04_2023_preview,
+            self._service_client_10_2021_dataplanepreview
+            if registry_name or registry_reference
+            else self._service_client_04_2023_preview,
             self._datastores,
             self._operation_container,
             requests_pipeline=self._requests_pipeline,
+            control_plane_client=self._service_client_04_2023_preview,
+            workspace_rg=self._ws_rg,
+            workspace_sub=self._ws_sub,
+            registry_reference=registry_reference,
             **app_insights_handler_kwargs,
         )
         self._operation_container.add(AzureMLResourceType.MODEL, self._models)
         self._code = CodeOperations(
-            self._operation_scope,
+            self._ws_operation_scope if registry_reference else self._operation_scope,
             self._operation_config,
             self._service_client_10_2021_dataplanepreview if registry_name else self._service_client_04_2023,
             self._datastores,
@@ -384,7 +442,7 @@ class MLClient:
         )
         self._operation_container.add(AzureMLResourceType.CODE, self._code)
         self._environments = EnvironmentOperations(
-            self._operation_scope,
+            self._ws_operation_scope if registry_reference else self._operation_scope,
             self._operation_config,
             self._service_client_10_2021_dataplanepreview if registry_name else self._service_client_04_2023_preview,
             self._operation_container,
@@ -416,7 +474,7 @@ class MLClient:
         self._operation_container.add(AzureMLResourceType.BATCH_ENDPOINT, self._batch_endpoints)
         self._operation_container.add(AzureMLResourceType.ONLINE_ENDPOINT, self._online_endpoints)
         self._online_deployments = OnlineDeploymentOperations(
-            self._operation_scope,
+            self._ws_operation_scope if registry_reference else self._operation_scope,
             self._operation_config,
             self._service_client_04_2023_preview,
             self._operation_container,
@@ -508,7 +566,10 @@ class MLClient:
         )
 
         self._featurestoreentities = FeatureStoreEntityOperations(
-            self._operation_scope, self._operation_config, self._service_client_06_2023_preview, **ops_kwargs
+            self._operation_scope,
+            self._operation_config,
+            self._service_client_04_2023_preview,
+            **ops_kwargs,
         )
 
         self._workspace_hubs = WorkspaceHubOperations(