+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_inputs_outputs/output.py
@@ -21,7 +21,7 @@ class Output(_InputOutputBase):
     """Define an output.
 
     :keyword type: The type of the data output. Accepted values are 'uri_folder', 'uri_file', 'mltable', 'mlflow_model',
-        'custom_model', and user-defined types.
+        'custom_model', and user-defined types. Defaults to 'uri_folder'.
     :paramtype type: str
     :keyword path: The remote path where the output should be stored.
     :paramtype path: Optional[str]
@@ -44,8 +44,7 @@ class Output(_InputOutputBase):
     :paramtype early_available: bool
     :keyword intellectual_property: Intellectual property associated with the output.
         It can be an instance of `IntellectualProperty` or a dictionary that will be used to create an instance.
-    :paramtype intellectual_property: Union[
-        ~azure.ai.ml.entities._assets.intellectual_property.IntellectualProperty,
+    :paramtype intellectual_property: Union[~azure.ai.ml.entities._assets.intellectual_property.IntellectualProperty,
         dict]
 
     .. admonition:: Example:
@@ -62,30 +61,11 @@ class Output(_InputOutputBase):
 
     @overload
     def __init__(self, type: Literal["uri_folder"] = "uri_folder", path=None, mode=None, description=None) -> None:
-        """Define a URI folder output.
-
-        :keyword type: The type of the data output. Can only be set to "uri_folder".
-        :paramtype type: str
-        :keyword path: The remote path where the output should be stored.
-        :paramtype path: str
-        :keyword mode: The access mode of the data output. Accepted values are
-            * 'rw_mount': Read-write mount the data
-            * 'upload': Upload the data from the compute target
-            * 'direct': Pass in the URI as a string
-        :paramtype mode: str
-        :keyword description: The description of the output.
-        :paramtype description: str
-        :keyword name: The name to be used to register the output as a Data or Model asset. A name can be set without
-            setting a version.
-        :paramtype name: str
-        :keyword version: The version used to register the output as a Data or Model asset. A version can be set only
-            when name is set.
-        :paramtype version: str
-        """
+        """"""
 
     @overload
     def __init__(self, type: Literal["uri_file"] = "uri_file", path=None, mode=None, description=None):
-        """Define a URI file outputs.
+        """Define a URI file output.
 
         :keyword type: The type of the data output. Can only be set to 'uri_file'.
         :paramtype type: str